rowCollapse {MatrixGenerics}	R Documentation

Extract one cell from each row (column) of a matrix-like object

Description

Extract one cell from each row (column) of a matrix-like object.

Usage

rowCollapse(x, idxs, rows = NULL, ...)

## S4 method for signature 'matrix_OR_array_OR_table_OR_numeric'
rowCollapse(x, idxs, rows = NULL, dim. = dim(x), ...)

## S4 method for signature 'ANY'
rowCollapse(x, idxs, rows = NULL, ...)

colCollapse(x, idxs = idxs, cols = NULL, ...)

## S4 method for signature 'matrix_OR_array_OR_table_OR_numeric'
colCollapse(x, idxs, cols = NULL, dim. = dim(x), ...)

## S4 method for signature 'ANY'
colCollapse(x, idxs = idxs, cols = NULL, ...)

Arguments

x
An NxK matrix-like object.

idxs
An index vector with the position to extract. It is recycled to match the number of rows (column)

rows, cols
A vector indicating the subset of rows (and/or columns) to operate over. If NULL, no subsetting is done.

...
Additional arguments passed to specific methods.

dim.
An integer vector of length two specifying the dimension of x, essential when x is a numeric vector. Note, that this is not a generic argument and not all methods need provide it.

Details

The S4 methods for x of type matrix, array, or numeric call matrixStats::rowCollapse / matrixStats::colCollapse.

Value

Returns a numeric vector of length N (K).

See Also

matrixStats::rowCollapse() and matrixStats::colCollapse() which are used when the input is a matrix or numeric vector.

Examples

  mat <- matrix(rnorm(15), nrow = 5, ncol = 3)
  mat[2, 1] <- NA
  mat[3, 3] <- Inf
  mat[4, 1] <- 0
  
  print(mat)
  
  rowCollapse(mat, idxs = 2)
  rowCollapse(mat, idxs = c(1,1,2,3,2))
  
  colCollapse (mat, idxs = 4)

[Package MatrixGenerics version 1.2.1 Index]