rowCounts {MatrixGenerics}	R Documentation

Count how often an element in a row (column) of a matrix-like object is equal to a value

Description

Count how often an element in a row (column) of a matrix-like object is equal to a value.

Usage

rowCounts(x, rows = NULL, cols = NULL, value = TRUE, na.rm = FALSE, ...)

## S4 method for signature 'matrix_OR_array_OR_table_OR_numeric'
rowCounts(x, rows = NULL,
  cols = NULL, value = TRUE, na.rm = FALSE, dim. = dim(x), ...)

## S4 method for signature 'ANY'
rowCounts(x, rows = NULL, cols = NULL, value = TRUE, na.rm = FALSE, ...)

colCounts(x, rows = NULL, cols = NULL, value = TRUE, na.rm = FALSE, ...)

## S4 method for signature 'matrix_OR_array_OR_table_OR_numeric'
colCounts(x, rows = NULL,
  cols = NULL, value = TRUE, na.rm = FALSE, dim. = dim(x), ...)

## S4 method for signature 'ANY'
colCounts(x, rows = NULL, cols = NULL, value = TRUE, na.rm = FALSE, ...)

Arguments

x
An NxK matrix-like object.

rows, cols
A vector indicating the subset of rows (and/or columns) to operate over. If NULL, no subsetting is done.

value
The value to search for.

na.rm
If TRUE, NAs are excluded first, otherwise not.

...
Additional arguments passed to specific methods.

dim.
An integer vector of length two specifying the dimension of x, essential when x is a numeric vector. Note, that this is not a generic argument and not all methods need provide it.

Details

The S4 methods for x of type matrix, array, or numeric call matrixStats::rowCounts / matrixStats::colCounts.

Value

Returns a integer vector of length N (K).

See Also

matrixStats::rowCounts() and matrixStats::colCounts() which are used when the input is a matrix or numeric vector.

For checks if any element is equal to a value, see rowAnys(). To check if all elements are equal, see rowAlls().

Examples

  mat <- matrix(rnorm(15), nrow = 5, ncol = 3)
  mat[2, 1] <- NA
  mat[3, 3] <- Inf
  mat[4, 1] <- 0
  
  print(mat)
  
  rowCounts(mat)
  colCounts(mat)
  
  rowCounts(mat, value = 0)
  colCounts(mat, value = Inf, na.rm = TRUE)

[Package MatrixGenerics version 1.2.1 Index]