rowMadDiffs {MatrixGenerics}	R Documentation

Calculates the mean absolute deviation of the difference between each element of a row (column) of a matrix-like object

Description

Calculates the mean absolute deviation of the difference between each element of a row (column) of a matrix-like object.

Usage

rowMadDiffs(x, rows = NULL, cols = NULL, na.rm = FALSE, diff = 1L,
  trim = 0, ..., useNames = TRUE)

## S4 method for signature 'matrix_OR_array_OR_table_OR_numeric'
rowMadDiffs(x, rows = NULL,
  cols = NULL, na.rm = FALSE, diff = 1L, trim = 0, ...,
  useNames = TRUE)

## S4 method for signature 'ANY'
rowMadDiffs(x, rows = NULL, cols = NULL, na.rm = FALSE,
  diff = 1L, trim = 0, ..., useNames = TRUE)

colMadDiffs(x, rows = NULL, cols = NULL, na.rm = FALSE, diff = 1L,
  trim = 0, ..., useNames = TRUE)

## S4 method for signature 'matrix_OR_array_OR_table_OR_numeric'
colMadDiffs(x, rows = NULL,
  cols = NULL, na.rm = FALSE, diff = 1L, trim = 0, ...,
  useNames = TRUE)

## S4 method for signature 'ANY'
colMadDiffs(x, rows = NULL, cols = NULL, na.rm = FALSE,
  diff = 1L, trim = 0, ..., useNames = TRUE)

Arguments

x
An NxK matrix-like object.

rows, cols
A vector indicating the subset of rows (and/or columns) to operate over. If NULL, no subsetting is done.

na.rm
If TRUE, missing values (NA or NaN) are omitted from the calculations.

diff
An integer specifying the order of difference.

trim
A double in [0,1/2] specifying the fraction of observations to be trimmed from each end of (sorted) x before estimation.

...
Additional arguments passed to specific methods.

useNames
If TRUE (default), names attributes of result are set. Else if FALSE, no naming support is done.

Details

The S4 methods for x of type matrix, array, table, or numeric call matrixStats::rowMadDiffs / matrixStats::colMadDiffs.

Value

Returns a numeric vector of length N (K).

See Also

matrixStats::rowMadDiffs() and matrixStats::colMadDiffs() which are used when the input is a matrix or numeric vector.

Examples

mat <- matrix(rnorm(15), nrow = 5, ncol = 3)
mat[2, 1] <- NA
mat[3, 3] <- Inf
mat[4, 1] <- 0

print(mat)

rowMadDiffs(mat)
colMadDiffs(mat)

[Package MatrixGenerics version 1.22.0 Index]