rowOrderStats {MatrixGenerics}	R Documentation

Calculates an order statistic for each row (column) of a matrix-like object

Description

Calculates an order statistic for each row (column) of a matrix-like object.

Usage

rowOrderStats(x, rows = NULL, cols = NULL, which, ...)

## S4 method for signature 'matrix_OR_array_OR_table_OR_numeric'
rowOrderStats(x, rows = NULL, cols = NULL, which, dim. = dim(x), ...)

## S4 method for signature 'ANY'
rowOrderStats(x, rows = NULL, cols = NULL, which, ...)

colOrderStats(x, rows = NULL, cols = NULL, which, ...)

## S4 method for signature 'matrix_OR_array_OR_table_OR_numeric'
colOrderStats(x, rows = NULL, cols = NULL, which, dim. = dim(x), ...)

## S4 method for signature 'ANY'
colOrderStats(x, rows = NULL, cols = NULL, which, ...)

Arguments

x
An NxK matrix-like object.

rows, cols
A vector indicating the subset of rows (and/or columns) to operate over. If NULL, no subsetting is done.

which
An integer index in [1,K] ([1,N]) indicating which order statistic to be returned

...
Additional arguments passed to specific methods.

dim.
An integer vector of length two specifying the dimension of x, essential when x is a numeric vector. Note, that this is not a generic argument and not all methods need provide it.

Details

The S4 methods for x of type matrix, array, or numeric call matrixStats::rowOrderStats / matrixStats::colOrderStats.

Value

Returns a numeric vector of length N (K).

See Also

matrixStats::rowOrderStats() and matrixStats::colOrderStats() which are used when the input is a matrix or numeric vector.

Examples

    mat <- matrix(rnorm(15), nrow = 5, ncol = 3)
    mat[2, 1] <- 2
    mat[3, 3] <- Inf
    mat[4, 1] <- 0
    
    print(mat)
    
    rowOrderStats(mat, which = 1)
    colOrderStats(mat, which = 3)

[Package MatrixGenerics version 1.2.1 Index]