rowQuantiles {MatrixGenerics}	R Documentation

Calculates quantiles for each row (column) of a matrix-like object

Description

Calculates quantiles for each row (column) of a matrix-like object.

Usage

rowQuantiles(x, rows = NULL, cols = NULL, probs = seq(from = 0, to = 1,
  by = 0.25), na.rm = FALSE, type = 7L, digits = 7L, ...,
  useNames = TRUE, drop = TRUE)

## S4 method for signature 'matrix_OR_array_OR_table_OR_numeric'
rowQuantiles(x, rows = NULL,
  cols = NULL, probs = seq(from = 0, to = 1, by = 0.25), na.rm = FALSE,
  type = 7L, digits = 7L, ..., useNames = TRUE, drop = TRUE)

## S4 method for signature 'ANY'
rowQuantiles(x, rows = NULL, cols = NULL,
  probs = seq(from = 0, to = 1, by = 0.25), na.rm = FALSE, type = 7L,
  digits = 7L, ..., useNames = TRUE, drop = TRUE)

colQuantiles(x, rows = NULL, cols = NULL, probs = seq(from = 0, to = 1,
  by = 0.25), na.rm = FALSE, type = 7L, digits = 7L, ...,
  useNames = TRUE, drop = TRUE)

## S4 method for signature 'matrix_OR_array_OR_table_OR_numeric'
colQuantiles(x, rows = NULL,
  cols = NULL, probs = seq(from = 0, to = 1, by = 0.25), na.rm = FALSE,
  type = 7L, digits = 7L, ..., useNames = TRUE, drop = TRUE)

## S4 method for signature 'ANY'
colQuantiles(x, rows = NULL, cols = NULL,
  probs = seq(from = 0, to = 1, by = 0.25), na.rm = FALSE, type = 7L,
  digits = 7L, ..., useNames = TRUE, drop = TRUE)

Arguments

x
An NxK matrix-like object.

rows, cols
A vector indicating the subset of rows (and/or columns) to operate over. If NULL, no subsetting is done.

probs
A numeric vector of J probabilities in [0, 1].

na.rm
If TRUE, missing values (NA or NaN) are omitted from the calculations.

type
An integer specifying the type of estimator. See stats::quantile() for more details.

digits
An integer specifying the precision of the formatted percentages. See stats::quantile() for more details.

...
Additional arguments passed to specific methods.

useNames
If TRUE (default), names attributes of result are set. Else if FALSE, no naming support is done.

drop
If TRUE a vector is returned if J == 1.

Details

The S4 methods for x of type matrix, array, table, or numeric call matrixStats::rowQuantiles / matrixStats::colQuantiles.

Value

a numeric NxJ (KxJ) matrix, where N (K) is the number of rows (columns) for which the J values are calculated.

See Also

matrixStats::rowQuantiles() and matrixStats::colQuantiles() which are used when the input is a matrix or numeric vector.

stats::quantile

Examples

mat <- matrix(rnorm(15), nrow = 5, ncol = 3)
mat[2, 1] <- NA
mat[3, 3] <- Inf
mat[4, 1] <- 0

print(mat)

rowQuantiles(mat)
colQuantiles(mat)

[Package MatrixGenerics version 1.22.0 Index]