rowAlls {MatrixGenerics}	R Documentation

Check if all elements in a row (column) of a matrix-like object are equal to a value

Description

Check if all elements in a row (column) of a matrix-like object are equal to a value.

Usage

rowAlls(x, rows = NULL, cols = NULL, value = TRUE, na.rm = FALSE, ...)

## S4 method for signature 'matrix_OR_array_OR_table_OR_numeric'
rowAlls(x, rows = NULL,
  cols = NULL, value = TRUE, na.rm = FALSE, dim. = dim(x), ...)

## S4 method for signature 'ANY'
rowAlls(x, rows = NULL, cols = NULL, value = TRUE, na.rm = FALSE, ...)

colAlls(x, rows = NULL, cols = NULL, value = TRUE, na.rm = FALSE, ...)

## S4 method for signature 'matrix_OR_array_OR_table_OR_numeric'
colAlls(x, rows = NULL,
  cols = NULL, value = TRUE, na.rm = FALSE, dim. = dim(x), ...)

## S4 method for signature 'ANY'
colAlls(x, rows = NULL, cols = NULL, value = TRUE, na.rm = FALSE, ...)

Arguments

x
An NxK matrix-like object.

rows, cols
A vector indicating the subset of rows (and/or columns) to operate over. If NULL, no subsetting is done.

value
The value to search for.

na.rm
If TRUE, NAs are excluded first, otherwise not.

...
Additional arguments passed to specific methods.

dim.
An integer vector of length two specifying the dimension of x, essential when x is a numeric vector. Note, that this is not a generic argument and not all methods need provide it.

Details

The S4 methods for x of type matrix, array, or numeric call matrixStats::rowAlls / matrixStats::colAlls.

Value

Returns a logical vector of length N (K).

See Also

matrixStats::rowAlls() and matrixStats::colAlls() which are used when the input is a matrix or numeric vector.

For checks if any element is equal to a value, see rowAnys().

base::all().

Examples

  mat <- matrix(rnorm(15), nrow = 5, ncol = 3)
  mat[2, 1] <- NA
  mat[3, 3] <- Inf
  mat[4, 1] <- 0
  
  print(mat)
  
  rowAlls(mat)
  colAlls(mat)
  

[Package MatrixGenerics version 1.2.1 Index]