rowAnyNAs {MatrixGenerics}	R Documentation

Check if any elements in a row (column) of a matrix-like object is missing

Description

Check if any elements in a row (column) of a matrix-like object is missing.

Usage

rowAnyNAs(x, rows = NULL, cols = NULL, ...)

## S4 method for signature 'matrix_OR_array_OR_table_OR_numeric'
rowAnyNAs(x, rows = NULL, cols = NULL, ...)

## S4 method for signature 'ANY'
rowAnyNAs(x, rows = NULL, cols = NULL, ...)

colAnyNAs(x, rows = NULL, cols = NULL, ...)

## S4 method for signature 'matrix_OR_array_OR_table_OR_numeric'
colAnyNAs(x, rows = NULL, cols = NULL, ...)

## S4 method for signature 'ANY'
colAnyNAs(x, rows = NULL, cols = NULL, ...)

Arguments

x
An NxK matrix-like object.

rows, cols
A vector indicating the subset of rows (and/or columns) to operate over. If NULL, no subsetting is done.

...
Additional arguments passed to specific methods.

Details

The S4 methods for x of type matrix, array, or numeric call matrixStats::rowAnyNAs / matrixStats::colAnyNAs.

Value

Returns a logical vector of length N (K).

See Also

matrixStats::rowAnyNAs() and matrixStats::colAnyNAs() which are used when the input is a matrix or numeric vector.

For checks if any element is equal to a value, see rowAnys().

base::is.na() and base::any().

Examples

  mat <- matrix(rnorm(15), nrow = 5, ncol = 3)
  mat[2, 1] <- NA
  mat[3, 3] <- Inf
  mat[4, 1] <- 0
  
  print(mat)
  
  rowAnyNAs(mat)
  colAnyNAs(mat)
  

[Package MatrixGenerics version 1.2.1 Index]