rowRanks {MatrixGenerics}	R Documentation

Calculates the rank of the elements for each row (column) of a matrix-like object

Description

Calculates the rank of the elements for each row (column) of a matrix-like object.

Usage

rowRanks(x, rows = NULL, cols = NULL, ties.method = c("max", "average"), ...)

## S4 method for signature 'matrix_OR_array_OR_table_OR_numeric'
rowRanks(x, rows = NULL,
  cols = NULL, ties.method = c("max", "average", "first", "last", "random",
  "max", "min", "dense"), dim. = dim(x), ...)

## S4 method for signature 'ANY'
rowRanks(x, rows = NULL, cols = NULL, ties.method = c("max", "average"), ...)

colRanks(x, rows = NULL, cols = NULL, ties.method = c("max", "average"), ...)

## S4 method for signature 'matrix_OR_array_OR_table_OR_numeric'
colRanks(x, rows = NULL,
  cols = NULL, ties.method = c("max", "average", "first", "last", "random",
  "max", "min", "dense"), dim. = dim(x), preserveShape = FALSE, ...)

## S4 method for signature 'ANY'
colRanks(x, rows = NULL, cols = NULL, ties.method = c("max", "average"), ...)

Arguments

x
An NxK matrix-like object.

rows, cols
A vector indicating the subset of rows (and/or columns) to operate over. If NULL, no subsetting is done.

ties.method
A character string specifying how ties are treated. Note that the default specifies fewer options than the original matrixStats package.

...
Additional arguments passed to specific methods.

dim.
An integer vector of length two specifying the dimension of x, essential when x is a numeric vector. Note, that this is not a generic argument and not all methods need provide it.

preserveShape
If TRUE the output matrix has the same shape as the input x. Note, that this is not a generic argument and not all implementation of this function have to provide it.

Details

The S4 methods for x of type matrix, array, or numeric call matrixStats::rowRanks / matrixStats::colRanks.

The matrixStats::rowRanks() function can handle a lot of different values for the ties.method argument. Users of the generic function should however only rely on max and average because the other ones are not guaranteed to be implemented:

max

for values with identical values the maximum rank is returned

average

for values with identical values the average of the ranks they cover is returned. Note, that in this case the return value is of type numeric.

Value

a matrix of type integer is returned unless ties.method = "average". Ithas dimensions' NxJ (KxJ) matrix, where N (K) is the number of rows (columns) of the input x.

See Also

matrixStats::rowRanks() and matrixStats::colRanks() which are used when the input is a matrix or numeric vector.

base::rank

Examples

  mat <- matrix(rnorm(15), nrow = 5, ncol = 3)
  mat[2, 1] <- NA
  mat[3, 3] <- Inf
  mat[4, 1] <- 0
  
  print(mat)
  
  rowRanks(mat)
  colRanks(mat)
  

[Package MatrixGenerics version 1.2.1 Index]